docker-weather-web-app/requirements.txt
Flask
requests
python-dotenv
pymongo
Flask-PyMongo
